refactor(resolve): split resolver modules
