+++ components/aihc-resolve/aihc-resolve.cabal
@@ -14,6 +14,11 @@ library
     Aihc.Resolve
     Aihc.Resolve.Types
 
+  other-modules:
+    Aihc.Resolve.Monad
+    Aihc.Resolve.Scope
+    Aihc.Resolve.Span
+
   hs-source-dirs: src
   build-depends:
     aihc-parser,
